rename histogram.py -> hist.py
