+++ astropy/visualization/hist.py
